// src/types/types_db.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export interface Database {
  public: {
    Tables: {
      bought_products: {
        Row: {
          created_at: string
          id: string
          metadata: Json | null
          price_id: string | null
          product_id: string | null
          sub_duration: string | null
          sub_type: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: string
          metadata?: Json | null
          price_id?: string | null
          product_id?: string | null
          sub_duration?: string | null
          sub_type?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: string
          metadata?: Json | null
          price_id?: string | null
          product_id?: string | null
          sub_duration?: string | null
          sub_type?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "bought_products_price_id_fkey"
            columns: ["price_id"]
            isOneToOne: false
            referencedRelation: "prices"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "bought_products_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "bought_products_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
      discord_usage: {
        Row: {
          created_at: string
          cycle_end: string | null
          cycle_start: string | null
          discord_id: string | null
          id: string
          usage: number | null
          voice_usage_characters: number | null
        }
        Insert: {
          created_at?: string
          cycle_end?: string | null
          cycle_start?: string | null
          discord_id?: string | null
          id?: string
          usage?: number | null
          voice_usage_characters?: number | null
        }
        Update: {
          created_at?: string
          cycle_end?: string | null
          cycle_start?: string | null
          discord_id?: string | null
          id?: string
          usage?: number | null
          voice_usage_characters?: number | null
        }
        Relationships: []
      }
      generated_messages: {
        Row: {
          created_at: string
          discord_id: string | null
          id: number
          language: string | null
          message: string | null
          sub_type: string | null
          variables: Json | null
        }
        Insert: {
          created_at?: string
          discord_id?: string | null
          id?: number
          language?: string | null
          message?: string | null
          sub_type?: string | null
          variables?: Json | null
        }
        Update: {
          created_at?: string
          discord_id?: string | null
          id?: number
          language?: string | null
          message?: string | null
          sub_type?: string | null
          variables?: Json | null
        }
        Relationships: []
      }
// ...
      settings: {
        Row: {
          created_at: string
          custom_slurs: Json | null
          eleven_labs_voice_id: string | null
          encrypted_eleven_labs_api_key: string | null
          id: string
          language: string | null
          league_aliases: Json | null
          league_name: string | null
          league_region: string | null
          league_tag: string | null
          twitch_bot_enabled: boolean | null
          updated_at: string | null
          user_id: string | null
          using_eleven_labs_default: boolean | null
        }
        Insert: {
          created_at?: string
          custom_slurs?: Json | null
          eleven_labs_voice_id?: string | null
          encrypted_eleven_labs_api_key?: string | null
          id?: string
          language?: string | null
          league_aliases?: Json | null
          league_name?: string | null
          league_region?: string | null
          league_tag?: string | null
          twitch_bot_enabled?: boolean | null
          updated_at?: string | null
          user_id?: string | null
          using_eleven_labs_default?: boolean | null
        }
        Update: {
          created_at?: string
          custom_slurs?: Json | null
          eleven_labs_voice_id?: string | null
          encrypted_eleven_labs_api_key?: string | null
          id?: string
          language?: string | null
          league_aliases?: Json | null
          league_name?: string | null
          league_region?: string | null
          league_tag?: string | null
          twitch_bot_enabled?: boolean | null
          updated_at?: string | null
          user_id?: string | null
          using_eleven_labs_default?: boolean | null
        }
        Relationships: [
          {
            foreignKeyName: "settings_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      subscriptions: {
        Row: {
          cancel_at: string | null
          cancel_at_period_end: boolean | null
          canceled_at: string | null
          created: string
          current_period_end: string
          current_period_start: string
          ended_at: string | null
          id: string
          metadata: Json | null
          price_id: string | null
          quantity: number | null
          status: Database["public"]["Enums"]["subscription_status"] | null
          trial_end: string | null
          trial_start: string | null
          user_id: string
        }
        Insert: {
          cancel_at?: string | null
          cancel_at_period_end?: boolean | null
          canceled_at?: string | null
          created?: string
          current_period_end?: string
          current_period_start?: string
          ended_at?: string | null
          id: string
          metadata?: Json | null
          price_id?: string | null
          quantity?: number | null
          status?: Database["public"]["Enums"]["subscription_status"] | null
          trial_end?: string | null
          trial_start?: string | null
          user_id: string
        }
        Update: {
          cancel_at?: string | null
          cancel_at_period_end?: boolean | null
          canceled_at?: string | null
          created?: string
          current_period_end?: string
          current_period_start?: string
          ended_at?: string | null
          id?: string
          metadata?: Json | null
          price_id?: string | null
          quantity?: number | null
          status?: Database["public"]["Enums"]["subscription_status"] | null
          trial_end?: string | null
          trial_start?: string | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "subscriptions_price_id_fkey"
            columns: ["price_id"]
            isOneToOne: false
            referencedRelation: "prices"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "subscriptions_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      users: {
        Row: {
          avatar_url: string | null
          billing_address: Json | null
          discord_id: string | null
          full_name: string | null
          id: string
          payment_method: Json | null
          twitch_channel: string | null
        }
        Insert: {
          avatar_url?: string | null
          billing_address?: Json | null
          discord_id?: string | null
          full_name?: string | null
          id: string
          payment_method?: Json | null
          twitch_channel?: string | null
        }
        Update: {
          avatar_url?: string | null
          billing_address?: Json | null
          discord_id?: string | null
          full_name?: string | null
          id?: string
          payment_method?: Json | null
          twitch_channel?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "users_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
// ...
  }
}
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
  ? Database["public"]["Enums"][PublicEnumNameOrOptions]
  : never